+++ src/js/handlers/components/getParams.mjs
@@ -1,0 +1,4 @@
+export function getQueryStringParam(param) {
+  const urlParams = new URLSearchParams(window.location.search);
+  return urlParams.get(param);
+}
