+++ src/js/templates/components/clearTemplate.mjs
@@ -1,0 +1,5 @@
+export function clearContainer(container) {
+  while (container.firstChild) {
+    container.removeChild(container.firstChild);
+  }
+}
