+++ src/js/templates/index.mjs
@@ -1,3 +1,4 @@
 export * from "./posts/post.mjs";
-export * from "./components.mjs";
 export * from "./profiles/profile-thumbnail.mjs";
+export * from "./components/clearTemplate.mjs";
+export * from "./components/message.mjs";
